switches lat/lng from float to double

+++ app/formats/json/TaskSubmissionFormats.scala
@@ -18,7 +18,7 @@ object TaskSubmissionFormats {
   case class AuditMissionProgress(missionId: Int, distanceProgress: Option[Float], completed: Boolean, auditTaskId: Option[Int], skipped: Boolean)
   case class AuditTaskSubmission(missionProgress: AuditMissionProgress, auditTask: TaskSubmission, labels: Seq[LabelSubmission], interactions: Seq[InteractionSubmission], environment: EnvironmentSubmission, incomplete: Option[IncompleteTaskSubmission], gsvPanoramas: Seq[GSVPanoramaSubmission], amtAssignmentId: Option[Int], userRouteId: Option[Int])
   case class AMTAssignmentCompletionSubmission(assignmentId: Int, completed: Option[Boolean])
-  case class LabelAccuracyPredictionSubmission(labelType: String, zoom: Int, severity: Option[Int], tagCount: Int, hasDescription: Boolean, lat: Option[Float], lng: Option[Float])
+  case class LabelAccuracyPredictionSubmission(labelType: String, zoom: Int, severity: Option[Int], tagCount: Int, hasDescription: Boolean, lat: Option[Double], lng: Option[Double])
 
   implicit val pointReads: Reads[Point] = (
     (JsPath \ "lat").read[Double] and
@@ -150,7 +150,7 @@ object TaskSubmissionFormats {
       (JsPath \ "severity").readNullable[Int] and
       (JsPath \ "tag_count").read[Int] and
       (JsPath \ "has_description").read[Boolean] and
-      (JsPath \ "lat").readNullable[Float] and
-      (JsPath \ "lng").readNullable[Float]
+      (JsPath \ "lat").readNullable[Double] and
+      (JsPath \ "lng").readNullable[Double]
     )(LabelAccuracyPredictionSubmission.apply _)
 }
